Fixed some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/535



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@3602 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/numfmtst.cpp
@@ -453,7 +453,7 @@ void NumberFormatTest::TestSecondaryGrouping(void) {
     CHECK(status, "createInstance(hi_IN)");
 
     UnicodeString out;
-    int32_t l = 1876543210L;
+    int32_t l = (int32_t)1876543210L;
     g->format(l, out);
     delete g;
     // expect "1,87,65,43,210", but with Hindi digits
